Address namespace collisions that will occur in cljs
Rewrite-clj has some namespace/fn combinations that will not work
under ClojureScript that need to be addressed in preparation
for ClojureScript support.

For example rewrite-clj.zip/find and rewrite-clj.zip.find collide
in ClojureScript.

Our strategy is:
- rename x.cljc to xz.cljc where xz was the name used in rewrite-cljs
- add x.clj, that simply proxies to xz.cljc, to restore compatibility
  with rewrite-clj

The exception is that rewrite-clj/node/coerce.clj simply gets renamed
to rewrite-cljs compatible coercer.cljc. This is an internal ns and
should not impact any rewrite-clj users.