Replace internal potemkin with cljc version
In preparation for ClojureScript support, we bring over
the cljc version of potemkin from rewrite-cljc dev efforts.

Potemkin's import-vars is controversial, but we are
starting with it for our initial rewrite-clj v1.

We complement import-vars with an import-vars-with-mods
which we use to bring in non-whitespace aware zipper fns
to the zip API.

We also carry on with rewrite-clj's use of potemkin's
defprotocol+ for clj. It is a passthrough to defprotocol
for cljs.