Preserve newline at end of input
When the last top level form is removed form a zipper, if that
form is followed by whitespace that includes 1 or more newlines,
a single newline is now preserved.

(It was kind of rude for rewrite-clj to strip newline at eof, so
it does not do that anymore)

Also reviewed rewrite-clj.zip/remove-preserve-newline, and updated
its docstring after better understanding it.

Closes #121