Address TODOs in node coercer - records
Portions of specialized handling for records not required.
Was added in innocence before understanding that all rewrite-clj
nodes need to coerce to themselves (they weren't at one point).

Specialized handling for records still needed for cljs.

Turfed incorrect comment that strongly implied order of
(extend-protocol ...) forms is important. It is not, added a reminder
comment as to why.