Appease, without understanding, shadow-cljs
I whacked this mole once before in rewrite-cljc-playground.

This time, perhaps the mole will stay down?

For some unknown (to me) reason, shadow-cljs will suffer a:
  #object[RangeError RangeError: Maximum call stack size exceeded]
when I use (thrown-with-msg? ...) in a custom zipper util tests.

I use thrown-with-msg? elsewhere without ill event, so?

Did not reproduce on macOS dev box using node 12.20.1 (which is bundled
with GraalVM at least when installed via skdman). Did reproduce with
node 15.9.0 on macOS dev box and node 14.5.5 on ubuntu up on
GitHub Actions).

Changes that triggered the error seem to be caused by changes
to other tests. So? Dunno.

For now have rewritten offending test to not use thrown-with-msg?