Deleted a commented out test
The test was verifying that if we were to z/insert-right* after a
comment node at end of input, rewrite-clj should make sure the inserted
node appeared on a newline.

This might be a bit of an unusual use case, as typically we skip
comments and newlines. Code could certainly work around the issue.

If it comes up organically from community can address.