Distingiush highlighting for lib test deps
I spit out deps for lib tests, its nice to easily eyeball that I am
testing with the versions of clojure, clojurescript and rewrite-clj
that I intend to.

To make it easier to focus on what I am interested in, I highlight
deps of interest. Now highlighting rewrite-clj and clojure deps
with different colours.