libs tests: tweak update lein deps skill lib
On a run on my dev box, I noticed the tests for this lib were failing
but the failure was not being detected by the libs tests script.

The recommended way to run tests for this lib always exits with 0,
even when tests fail.

We now run the tests a second time via node, which will exit with
non-zero on failure.

The reason this test was failing on my dev box was because I had
selected GraalVM via SDKman as my java. This includes a version of
node under GraalVM which caused the failure.