dev: Update babashka scripts to avoid side effects
This might be considered a bit overly meticulous at this point but it
could save confusion at some later date.

Scripts were previously always invoking the main function.
Thing is, this is not always technically appropriate.
For example, if someone issues a `bb doc` on a rewrite-clj script var,
main will be invoked.

With a pointer from borkdude to check out
https://book.babashka.org/#main_file, we now only automatically invoke
main when the clj file was invoked as a script.

I also adjusted all scripts to have a proper `-main` so that
for example:
./script/apply_import_vars.clj check
can alternatively be invoked via:
bb -m apply-import-vars check

I'll stick with the ./script invocation in our docs, as most shells
offer autocompletion for files.