tests: Update GraalVM to 21.1.0
Drop native image testing for JDK8. It is no longer an option
for macOS and we did not test on Windows due to complexity of setup.
So we now only currently test native image for JDK11.

Drop native-image --enable-all-security-services, no longer
needed and deprecated.

Update fail fast check for Graal version. The native-image cmd changed
text format for --version so we adapt.