dev: update script/task deps, use docopt for args
After upgrading docopt, decided to take a crack at using it for all
task command line args for consistency for users and for
maintainability. I was previously using a mix of adhoc, docopt and
clojure.tools.cli.

I think overall, this went quite well. I wrote a little wrapper-hack
helper.main/doc-arg-opt to allow us to specify "Valid args: ..."
instead of "Usage: my-exe...". The Usage syntax focuses on the
executable, which is really not appropriate for tasks, so we focus
on the args only.

If docopt also supported validation of option arguments
(ex. --my-option OPTION_ARG) that'd be awesome, but it doesn't so
had to include extra validation for these args.