Git ignore clj-kondo rewrite-clj export config
When clj-kondo is asked to copy configs it finds on the classpath
to the .clj-kondo dir, our local export rewrite-clj config will
sometimes be picked up.

This export config is for users of rewrite-clj, not for development of
rewrite-clj itself. Easiest thing to do is understand that it will
happen and have git ignore it.