lint: add camsaul's whitespace-linter
While working on rewrite-clj, I've tried a variety of IDEs and editors.
Some of them were not great a trimming trailing whitespace.
Or maybe I was not great a configuring them appropriately.

When making a change, I don't want to deal with spurious diffs
due to trailing whitespace. For this reason, I've added camsaul's
whitespace-linter to our lint task.

One could argue that we should instead add cljfmt to our linting
to catch a larger variety of common whitespace diffs. Maybe some day
we'll do that, but right now I think camsaul's linter is a gentler
compromise.