Graal native-image no longer fails with graal-build-time
Because the Graal team has deprecated use of global
--initalize-at-build-time for native-image, we want
to switch to using clj-easy/graal-build-time.

After attempting this switch, our Windows native-image
builds started to consistently to fail.

The resolution turned out to be simple and has allowed
us to re-apply clj-easy/graal-build-time.

Details:

The native-image -H:Name option is used to specify a
target executable.

It seems that on Windows, some components of native-image
recognize that -H:Name can include a path and others do not.
And/or some components understand a forward slash path
separator in -H:Name and others do not.

We do not need to work out the nuances of what paths
and syntax work for -H:Name because we now instead
specify our target path via -H:Path.

We can assume that switching away from a global
--initialize-at-build-time evokes different components
in GraalVM native-image. It might be interesting to
explore why someday, but that is out of scope for this
issue.

Closes #165