Fix end position on root node of untracked zipper
The end position of the root node of a parsed string was inherited from
the first child. This meant that the root node looked like it spanned
only one child, instead of all of its children. This fix makes the root
node take its end position from its last child.

After this fix, the end position of an untracked zipper (which uses the
parser's data) matches the end position of a tracked zipper (which uses
node extents).