Restore single pass over form children
This reverts to a scenario where parse-all does only a single pass over
the children nodes.

It does this by replacing a `doall` + `last` with just `last`. The
`last` is new code as of the prior commit, so this switches back to a
single pass over the children. It's not clear what the `doall` was
doing. Perhaps it ensured that a file was fully parsed before closing,
though I tried and failed to confirm that. In any case, the `last`
should serve the same purpose.