internal deps: bump shadow-cljs
Now explicitly depending on test.check for shadow-cljs test, formerly
got it only because it happened to be on the shadow-cljs classpath.
With this newest version this seems to be no longer the case.

The shadow-cljs.edn config file formerly only lived for the life of the
shadow-cljs test run.
This meant less chuff in the project root and less duplication of code,
but also made troubleshooting more difficult.
Now have an explicit tracked shadow-cljs.edn.