ci: run canary tests in matrix one job per lib
Take a stab at running each canary lib test as a GitHub Action job.

If this works, it should run canary tests in parallel, which is nice.

But the real benefit will be able to rerun a single failed canary lib
that has flaked out on CI rather than having to rerun them all.