Move to release-num versioning scheme
For 1.0 we were using a 1.0.<commit-count> version scheme.
Moving forward we'll use the simpler (at least for me) 1.1.<release-num>
version scheme.

We start <release-num> at 44 to account for the 44 known existing
releases of rewrite-clj. Next release will be 1.1.45.

There is some script code around built-versions that could be simplified.
This was determined by the build in the past, but we know what our
built-version will be prior to our build now.
But I've left it as is for now.

Added a new build dir for versioning code. This code is not entirely
shared by build.clj and babashka, but having everything that talks about
versions in one place makes good sense.

Closes #179