maint: hack: use focal planck on jammy
There are no Ubuntu jammy planck binaries yet.

For now we'll hack around this by using the Ubuntu focal binaries for
planck on yammy.

That last official binaries built for planck on ubuntu are v2.25.0.
So we are behind the current planck version of v2.27.0.
But that's probably fine for now.