Support Clojure 1.12 array class tokens
The new `foo/3` was reader-illegal prior to 1.12.

Bring in and customize some tools.reader bits and bobs to support
reading this new syntax.

Clojure's `symbol` allows us to construct technically illegal symbols,
we take advantage of this to test and support this new syntax in any
version of Clojure via, for example, `(symbol "foo" "3")`.

Closes #279