decorator nodes: wrap around - first pass
Always wrap decorated node. Not sure if this is the best way to go.
If on or in the decoration, we wrap the decorated node:

  |'a -> [|'a]
  '|'a -> ['|'a]

But if at the undecorated node, we could wrap at that level:

  '|a -> '[|a]

But as of this commit we always wrap at the root of the decoration:

  '|a -> ['|a]