test & dev: bump deps (#380)
* test: bump shadow-cljs and clojurescript

These seem to work fine with rewrite-clj locally, the buzz in the chats,
is there might be more changes coming...

* dev: bump deps

* test libs: bump cljfmt, clojure-lsp

* ci: adapt: new min jdk for cljs/shadow-cljs

Allow test-doc to run only clj or cljs tests.

cljs tests now run on jdk11 (was jdk8)
shadow-cljs tests now run on jdk21 (was jdk11)

* ci: lint with latest jdk

Was linting with earliest supported jdk, but linting with latest is
better because eastwood will find and report deprecations in latest jdk.

* test: bump shadow-cljs to 3.0.3

* ci: use jdk11 for planck cljs-boostrap tests

Was jdk8 but Clojurescript now requires jdk11.